+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 2.2.0
+- Added `materialize()` method on `NSManagedObject`.
+- `Collection.fetchFaultedObjects()` deprecated, use `Collection.materializeFaultedObjects()` instead.
+- More tests.
+
 ### 2.1.1
 - Minor refinements.